BACnetEDE.ObjectTypes (bacnet_ede v0.1.1)

The Object Types contain a number of rows or data, identified by its reference number, listing an enumeration of BACnet object types. BACnet object types are defined by the ASHRAE 135 BACnet protocol standard.

The library bacstack contains BACnet protocol enumerations, so if you already use a BACnet library, you do not need to parse object types files, because the object types are standardized in the BACnet protocol. Only proprietary object types are not standardized (which no one should ever use).

Types

dump_option()

@type dump_option() :: {:csv_encoder, module()} | {:fill_all_columns, boolean()}

Available options for to_binary/2, to_file/3 and to_stream/2.

See dump_options/0 for a description of the available options.

dump_options()

@type dump_options() :: [dump_option()]

Available options for to_binary/2, to_file/3 and to_stream/2.

Available options:

csv_encoder: module() - Optional. A module implementing the Nimble CSV behaviour. The default CSV encoder uses semicolons and CRLF newlines as separator. You may have the need for a different separator (i.e. comma) for your file.

fill_all_columns: boolean() - Optional. Ensures all rows have the same amount of columns. By default, only the necessary amount of columns per row is used for the header part.

parse_option()

@type parse_option() :: {:csv_parser, module()}

Available options for from_binary/2, from_file/2 and from_stream/2.

See parse_options/0 for a description of the available options.

parse_options()

@type parse_options() :: [parse_option()]

Available options for from_binary/2, from_file/2 and from_stream/2.

Available options:

csv_parser: module() - Optional. A module implementing the Nimble CSV behaviour. The default CSV parser uses semicolons and CRLF newlines as separator. You may have the need for a different separator (i.e. comma) for your file.

t()

@type t() :: %BACnetEDE.ObjectTypes{
  object_types: %{optional(non_neg_integer()) => String.t()}
}

Object Types are keyed by the BACnet protocol enumeration number to the text as specified in the EDE data (something like "Analog Input").

The text is though transformed from underscored to a space separated title case. For example analog_input is changed to Analog Input, analog output to Analog Output. But no further modifications are done.

Functions

from_binary(binary, opts \\ [])

@spec from_binary(binary(), parse_options()) :: {:ok, t()} | {:error, term()}

Parses the given binary as an Object Types EDE file.

See parse_options/0 for a description of the available options.

from_file(path, opts \\ [])

@spec from_file(Path.t(), parse_options()) :: {:ok, t()} | {:error, term()}

Parses an Object Types EDE file from a CSV file path. This function uses line streaming for the parsing process.

See parse_options/0 for a description of the available options.

from_stream(stream, opts \\ [])

@spec from_stream(Enumerable.t(), parse_options()) :: {:ok, t()} | {:error, term()}

Parses an Object Types EDE file from a CSV stream (i.e. a file stream).

Make sure the stream is line-orientated or use NimbleCSV's to_line_stream/1 if you can't. See BACnetEDE.CSV.to_line_stream/1 for the default parser's function.

See parse_options/0 for a description of the available options.

to_binary(object_types, opts \\ [])

@spec to_binary(t(), dump_options()) :: {:ok, binary()} | {:error, term()}

Dumps the given Object Types EDE data into a CSV encoded binary.

See dump_options/0 for a description of the available options.

to_file(object_types, path, opts \\ [])

@spec to_file(t(), Path.t(), dump_options()) ::
  {:ok, Enumerable.t()} | {:error, term()}

Dumps the given Object Types EDE data into a CSV encoded file.

See dump_options/0 for a description of the available options.

to_stream(object_types, opts \\ [])

@spec to_stream(t(), dump_options()) :: {:ok, Enumerable.t()} | {:error, term()}

Dumps the given Object Types EDE data into a CSV encoded stream.

See dump_options/0 for a description of the available options.